server/serialConnection/requirements.txt
websocket_client
pyee
pyserial
